api/pyproject.toml
[project]
name = "doccelerate-api"
version = "0.1.0"
description = "Doccelerate API"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi",
    "uvicorn",
    "python-dotenv",
    "pydantic",
    "pydantic-settings",
    "pydantic-core",
]

